+++ src/test/java/com/structurizr/dsl/InlineScriptDslContextTests.java
@@ -2,6 +2,8 @@ package com.structurizr.dsl;
 
 import org.junit.jupiter.api.Test;
 
+import java.io.File;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
@@ -10,7 +12,7 @@ class InlineScriptDslContextTests extends AbstractTests {
     @Test
     void test_end_ThrowsAnException_WhenAnUnsupportedLanguageIsSpecified() {
         try {
-            InlineScriptDslContext context = new InlineScriptDslContext(new WorkspaceDslContext(), "java");
+            InlineScriptDslContext context = new InlineScriptDslContext(new WorkspaceDslContext(), new File("workspace.dsl"), "java");
             context.end();
             fail();
         } catch (Exception e) {